pyproject.toml
[tool.poetry]
name = "online-trading-platform"
version = "0.1.0"
description = ""
authors = ["Yuhen Liubimov <yauhenliubimau@gmail.com>"]
readme = "README.md"
packages = [{include = "online_trading_platform"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
celery = {extras = ["redis"], version = "^5.3.6"}
environs = "^10.0.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
flake8 = "^6.1.0"
isort = "^5.13.2"
mypy = "^1.7.1"

[tool.mypy]
cache_dir = ".venv/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = true
exclude = [
    ".venv",
    ".github",
    ".idea",
    ".vscode",
    "venv",
]
files = ["."]
pretty = true
python_version = "3.10"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true



[tool.black]
line-length = 79
target-version = ["py310"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.local.*
    | .*\.venv.*
    | .*\.idea.*
    | .*\.vscode.*
)
'''

[tool.isort]
force_single_line = true
line_length = 79
multi_line_output = 3
profile = "black"
skip = [
    ".venv",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
